moved the cli command to its own folder for better comprehension
